Add playGame function and add more functions
The playGame function plays multiple rounds of rock, paper, scissors
until one of the players reaches 3 points.

Multiple new functions had to be added within the playGame function to
determine if a new round needed to be triggered or if the game was over.

I also updated the logic to determine if one of the players got to three
points, rather than if the total score between both was five, to be the
new end-game trigger.

This commit is not as atomic as I would like because there was a lot of
trial and error to get this to work. I want to learn how to be better
at commiting even when I have to do trial and error for many portions
to get something working.