Add back in code to determine overall winner
Updated the code so that it was looking to determine a winner based on
who got to 5 points first and then delete the prompt and buttons once
the player or the computer has won.

Added a prompt div in the HTML file as well to explain the point to the
player.

I also deleted out all the unused code from the previous version before
there was a UI.